map_manager.py file created
